ref: renaming, graphs and planning
