+++ lib/pages/graphs/widgets/categories/categories_bar_chart.dart
