+++ lib/pages/graphs/widgets/linear_progress_bar.dart
