+++ lib/pages/transactions/widgets/month_selector.dart
@@ -7,7 +7,7 @@ import '../../../providers/transactions_provider.dart';
 import '../../../ui/device.dart';
 import '../../../ui/extensions.dart';
 import '../../../ui/formatters/formatted_date_range.dart';
-import '../../graphs_page/widgets/categories/categories_bar_chart.dart';
+import '../../graphs/widgets/categories/categories_bar_chart.dart';
 
 enum MonthSelectorType { simple, advanced } //advanced = with amount
